Addresses peer review comments
